+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-NOTE: we need to add changed carrid over from 0.7.2 when finished
+NOTE: we need to add changes carried over from 0.7.2 when finished
 
 ## [0.9.0]
 ### Added
@@ -9,6 +9,7 @@ NOTE: we need to add changed carrid over from 0.7.2 when finished
 
 ### Changed
  - API BREAKING CHANGE: cmd api has been removed
+ - command line norst is made default (deprecated) and command forcerst added to force reset of a sensor at first connect
 
 ## [0.8.x]
 ### Refer to CHANGELOG in the [v.0.8.x] branch